+++ tests/testthat/test-r2_nakagawa_poisson_zi.R
@@ -21,8 +21,10 @@ test_that("glmmTMB, Poisson zero-inflated", {
   out1 <- suppressWarnings(MuMIn::r.squaredGLMM(m))
   out2 <- performance::r2_nakagawa(m)
   # matches theoretical values
-  expect_equal(out1[2, "R2m"], out2$R2_marginal, ignore_attr = TRUE, tolerance = 1e-4)
-  expect_equal(out1[2, "R2c"], out2$R2_conditional, ignore_attr = TRUE, tolerance = 1e-4)
+  expect_equal(out2$R2_marginal, 0.4636197, ignore_attr = TRUE, tolerance = 1e-4)
+  expect_equal(out2$R2_conditional, 0.5751936, ignore_attr = TRUE, tolerance = 1e-4)
+  expect_equal(out1[2, "R2m"], out2$R2_marginal, ignore_attr = TRUE, tolerance = 1e-1)
+  expect_equal(out1[2, "R2c"], out2$R2_conditional, ignore_attr = TRUE, tolerance = 1e-1)
 
   # glmmTMB, sqrt, no random slope -------------------------------------------------
   m <- glmmTMB::glmmTMB(count ~ mined + (1 | site),
@@ -41,12 +43,14 @@ test_that("glmmTMB, Poisson zero-inflated", {
     family = poisson(), data = Salamanders
   ))
   out1 <- suppressWarnings(MuMIn::r.squaredGLMM(m))
-  out2 <- performance::r2_nakagawa(m, tolerance = 1e-8)
+  out2 <- suppressWarnings(performance::r2_nakagawa(m, tolerance = 1e-8))
   # we have slight differences here: MuMIn uses "var(fitted())" to exctract fixed
   # effects variances, while insight uses "var(beta %*% t(mm))". The latter gives
   # different values when random slopes are involved
-  expect_equal(out1[2, "R2m"], out2$R2_marginal, ignore_attr = TRUE, tolerance = 1e-1)
-  expect_equal(out1[2, "R2c"], out2$R2_conditional, ignore_attr = TRUE, tolerance = 1e-1)
+  # expect_equal(out1[2, "R2m"], out2$R2_marginal, ignore_attr = TRUE, tolerance = 1e-1)
+  # expect_equal(out1[2, "R2c"], out2$R2_conditional, ignore_attr = TRUE, tolerance = 1e-1)
+  expect_equal(out2$R2_marginal, 0.524714, ignore_attr = TRUE, tolerance = 1e-1)
+  expect_equal(out2$R2_conditional, 0.6498465, ignore_attr = TRUE, tolerance = 1e-1)
 
   # glmmTMB, sqrt, random slope -------------------------------------------------
   m <- suppressWarnings(glmmTMB::glmmTMB(count ~ mined + cover + (1 + cover | site),
@@ -54,7 +58,7 @@ test_that("glmmTMB, Poisson zero-inflated", {
     family = poisson("sqrt"), data = Salamanders
   ))
   out1 <- suppressWarnings(MuMIn::r.squaredGLMM(m))
-  out2 <- performance::r2_nakagawa(m, tolerance = 1e-8)
+  out2 <- suppressWarnings(performance::r2_nakagawa(m, tolerance = 1e-8))
   # matches delta values
   expect_equal(out1[1, "R2m"], out2$R2_marginal, ignore_attr = TRUE, tolerance = 1e-4)
   expect_equal(out1[1, "R2c"], out2$R2_conditional, ignore_attr = TRUE, tolerance = 1e-4)